LPS-44098 Start by removing the lines in DLFileEntryServiceImpl causing the regression bug
This will fix some errors due to portal thinking that files are locked when in fact they are not. To keep the override checkout functionality I move the checks to the upper levels. Note that this change will remove the 'Cancel checkout' button from the UI but we will add it later.