LPS-46822 when deleting a folder, delete all files and their related elements (assets, etc.)
In the deleteFolder method I need to call the deleteFolder method on the repository before any
call to dlAppHelper; otherwise, a user without permissions in the backend repository would delete
part of the information (assets, sync events would be thrown, etc).

This makes the ordering of statements a little weird, as I need to get all file entries and the
folder itself before calling the repository.