LPS-47269 trigger the event in all methods that called the `hasLockedEntry`
For all methods that called directly or indirectly the `hasLockedEntry` method,
insert a call to the previously extracted method.

I've ignored some methods where it was obviously wrong to generate a Sync event
(e.g. getFile() and other read-only methods that should not trigger an update event)
but that called the `hasLockedEntry` method indirectly.