LPS-47508 add a `TrashCapability` to move trash logic into repositories
Currently, only the LiferayRepository supports the trash bin, but it does make
sense for other kind of repositories to support it. By representing the Trash
functionality with a capability we avoid polluting the repository interfaces
with methods that only a small subset of repositories actually need, still
allowing those that need it to implement the capability.