LPS-50073 Documents in the 'Home' root folder are not searchable
The problem was caused by DLFileEntryIndexer. In prior versions of
Liferay, files in the root folder used to have treePath field of Lucene
documents set to 0 (that was incoherent with files stored outside the
root folder, but such is life).

What I'm doing in this commit is setting the value of the field to 0.
It's not the cleaner solution but it helps backward compatibility.