LPS-51191 test WF when adding a file entry
For testing the workflow I introduce a testing pattern based on the RAII
idiom of C++.

In some tests there is a need to slightly modify the environment (in
this particular case, I wanted to intercept calls to WF
handlers to count the number of invocations). Making sure that the
env. is modified and cleaned up after the test ends using the common
Java idioms (try / finally) lends to very verbose tests. It's also not
possible to move this logic into a JUnit listener, as it would be very
expensive to always instrument classes for all tests (and doesn't make a
lot of sense unless really needed).

I've taken advantage of Java 7 try-with-resources syntax for this
purpose. The idea is to encapsulate this environment changes in an
object, where the state is initialized in the constructor and cleaned up
in the `close` method. The try block semantics will guarantee that
everything works correctly and all the logic will be encapsulated in an
object (making it reusable across different tests and composable).