LPS-51191 make DLApp*ServiceTest pass
The file version to use when changing status will need to be fetched
before or after the call to update/revert depending on the
circumstances. This is so because inside de update/revert method an auto
check in/out may be made, changing the version under the hood. To handle
this we use FileVersionRefernece, which is responsible of reading the
correct file version at the correct moment. Note that once we move the
check in code into a capability (in a couple of commits), this class
won't be needed anymore and will be removed.

Also, we need to refresh the FileEntry before returning it. Otherwise,
the version changes won't be propagated correctly.