LPS-53640 Provide taglib as a workaround
BND infers that the module requires the productivity-center tag, but it
is not able to guess that the taglib itself is contained within the
module. So, when deploying the web module, the dependency is not
resolved and the module sits there in a "installed" states.

As a workaround, declare the taglib as a provided capability.