LPS-56056 Handle all user errors
There are 2 exceptions that I've left unhandled, and thus will produce a
fatal error:

  - EntrySmallImageNameException
  - EntrySmallImageSizeException

There are raised only when an internal inconsistency occurs (the
name/size of the image doesn't match the underlying file entry ones),
and should never occur. Also, if that situation does occur, the user has
nothing to do, so it's better to left it unhandled.