LPS-66628 Make sure dependent items get refreshed
The current cache implementation is too "dumb", so we have no reliable
way to clear file versions when its file entry is invalidated. We're
studying better approaches to solve this, but for the moment the only
low-risk option is to have a pessimistic cache policy, cleaning it out
completely when necessary.