LPS-67515 Handle external repositories correctly
In some places, the code assumed that any repository that was not the
default one was external. Also, for the PortletFileRepository, it was
not importing the files.

I've changed that so that no "internal" repositories are ignored, no
hardcoded references to repository classNames are used, and file from
the PortletFileRepository are correctly imported.

This last point may require changes in StagedModelDataHandlers for
entities that include attachments or additional content (e.g. Cover
images in blogs), as right now those files are imported in an ad-hoc
fashion in each implementation. As the LAR will now include those files,
they will be imported automatically, so any special logic needs to be
removed.