LPS-67515 Save version UUID in the LAR
While importing, the version UUID is required, but it was not available
in the LAR. Also, to get it, it relied on DL services, so it was
impossible to import a LAR file that contained files which didn't exist
in the database.

To fix this, I added a new XML file to contain the required version
metadata (right now, only the UUID). We should check with the staging
guys, in case there is another (better) approach to add extra
information.