LPS-67515 Get data from the LAR, not from the DB
The same as before: this code assumed the file entry that was being
imported already existed in the database.

Note that this fix is not entirely correct, as the cover image is
already being imported by the file entry handler. I think a proper fix
for this would be to get the new file entry ID from the keys and update
the blogs entry (in pseudo code):

  map = portletDataContext.getNewPrimaryKeysMap(FileEntry.class);

  correctFileEntryId = map.get(oldFileEntryId);

  importedBlogEntry.setCoverImageFileEntryId(correctFileEntryId);

But for the moment, as I didn't have time to test this thoroughly, I'm
keeping this as is so that the Frontend team can resume working (this is
a blocking issue for them). The main consequence is that cover images
are imported twice: once by the file entry data handler, and another
time by the blogs entry data handler. We should let the file entry data
handler do the work and remove this logic from here.