LPS-68339 Rename Article portlet repository
Older versions of KB used a different repository for the Article portlet
than for the rest of KB applications. The bug was fixed, but this wasn't
taken into account when writing the upgrade processes.

Create a new upgrade step to make sure all KB repositories point to the
main one.