LPS-67974 Override current URL attribute
When this code is reached, the optimal thing to do would be to do an
HTTP redirect to the main D&M view, displaying an error message. But
there are two problems here:

  - Doing an HTTP redirect during the render phase seems to be not
    allowed by the portlet spec.
  - Even if we do a redirect, the session error gets cleaned before
    reaching the liferay:error tag.

The redirect is needed so that the `folderId` parameter that causes the
ping failure is not used by the JSP. Otherwise, almost all action URLs
will receive a `redirect` parameter with the original URL, which will
fail after the redirect.

By replacing the currentURL we get around this problem, as all URLs with
a `redirect` parameter will point to the correct place. Also, note that
technically this is not incorrect, as rendering `view.jsp` with no
folder (this case) is functionally equivalent to the redirect.