LPS-68127 Use server side rendering to avoid CSS trickery
While the CSS solution is way cleaner than this one, it is too
brittle: adding any content (even newlines or any other blank) to the
`small' element will cause the ` - ' not to be rendered.

This was the reason why the formatting of the `c:if' was weird (see
previous SF commit); by not breaking the line before the `c:if', no
spurious blanks were inserted and everything worked fine.