LPS-69656 Remove only already existing resources
Previously, if there was already any resource action for the new
portlet, it was assumed that the upgrade had finished with the
resource actions. In the general case, this should be true, but the
Announcements portlets had different resources depending on the
version (SO or portal), and so some resource actions were lost in the
process.