LPS-69656 Add new upgrade step
To make sure all users get upgraded correctly, we need to guarantee
that both upgrade processes (upgradePortletId and upgradePermission)
are executed no matter what.

As UpgradePermission is not idempotent (i.e. it cannot be executed
more that once), we need to add an optional parameter to customize
error logging and ignore the case where the upgrade process has been
already executed.