LPS-70180 Fix variables in templates
There may be some problem with the baseline tool. It is detecting API
changes where there seems to be none. In this particular case:

  - The superclass (SubscriptionSender) has (among others) a public
    `void initialize(void)` and a protected `String
    replaceContent(String, Locale)` methods. Obviously, both methods
    are part of the subclass API.

  - The subclass overrides the `String replaceContent(String, Locale)`
    method.

After moving some code around, the situation changed to the following:

  - The subclass overrides the `void initialize(void)` method.

  - The subclass no longer overrides the `replaceContent` method.

No changes to the superclass were made, and the set of methods exposed
by the subclass is exactly the same as before (i.e. no API changes),
but `gw baseline` is complaining about incompatible changes.