LPS-70043 Supporting config for all contexts
A config line like:

    general=mentions,categorization

Will now not only apply to form navigators without a context provider
component associated to them but also to those with a context that has
no specific config set for them.

The user will be able to override that config for a specific context by
adding a line with context like this:

    add.general=mentions

Which will only apply to the "add" context. The original config line
will still apply to all of the other contexts.