LPS-70675 Use local service call
When Guest is replying to a message, the addMessage call succeedes, as
she has the REPLY_TO permission. Problem is, the update call performed
just after that fails, as it requires the UPDATE permission.

Use the local service, as if the user had no permission, the add call
would have failed.

Also, note that we cannot move this logic into the service, as this is
something that is only necessary for the portlet because of the way
drag & drop attachments work. Regular (remote) service calls should
not need to do this.