LPS-70620 Create handler for content plugged in from outside D&M
There is some content that has no explicit references in staging,
either because it has no model (e.g. previews or thumbnails) or
because it is a feature "core" D&M has no knowledge of (e.g. Adaptive
Media).

This extension point will allow features that depend on D&M to work
properly with staging.