LPS-71420 Make dependency less strict
It should be possible to override the content processor -- in fact, we
need to do that for Adaptive Media to work -- but the dependency is too
specific (it depends on a non exported class), and being reluctant it
cannot be overriden at runtime.

Make it depend on the interface instead of on a particular implementation,
and let the policy option be "greedy", so that new components with higher
ranking will get wired in.