LPS-73365 Don't use a dummy proxy
The problem with the dummy proxy is that it will return null when
asked about any capability, so repositories would be incorrectly
initialized *and* cached by the D&M infrastructure.

This could happen if any user tries to do something D&M or file
related in the portal just as it finishes initializing. There's a very
short time window in which some D&M components may not be fully
enabled, and any usage of the repositories would leave D&M unusable.

This commit won't fix the problem -- for that we need to extract stuff
out of the core -- but at least it will avoid caching corrupted
repositories, and it will fail immediatelly, making it easier to
localize the error.

In parallel, we'll start extracting D&M components out of the core to
fix this; we'll do that under different tickets.