LPS-73824 Add the concept of builder to demo data creators
Builder will behave like any other builder that follows the builder pattern. Instances of those classes will be used to hold information about the operations that developers want to perform on a model created by a demo data creator, such as adding comments to a BlogsEntry.

Demo data creators that support extra operations, will have a new method `newBuilder` that will return a new instance of a builder for that demo data creator.