LPS-84868 Introduce new type to accuratelly represent version changes
Currently, a boolean is used as a parameter to methods that may
potentially bump a file version number. This is not correct, as there
are 3 possible states: minor version, major version or none.

Previously, there was logic in place to guess what the user
wanted (removed in LPS-82101). This new type will let callers specify
explicitly what they really want.