LPS-84735 Not needed
This check only serves to save the message from being sent at the
expense of adding a method to the public API. The same check is also
being performed in the message listener when the autoTagger is called so
we don't lose any functionality.

Also in the case the check were true it would be called twice so we are
saving one call by removing it.